ich fick dein mutta

+++ src/main/java/tn/esprit/dari/repositories/SubscribeRepository.java
@@ -3,7 +3,7 @@ package tn.esprit.dari.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.dari.entities.Subscribe;
 
-public interface SubscribeRepository extends JpaRepository<Subscribe, Long> {
+public interface SubscribeRepository extends JpaRepository<Subscribe, Integer> {
 
 
 }
